requirements.txt
# sudo apt install ffmpeg
google-api-python-client==2.52.0
google-auth-oauthlib==0.5.2